Added dependency checking to genesis.
Starting in v1.5.0, genesis will check for
a `.genesis_deps` file in the deployment root.
If found, it parses it to find commands that should
be at least version X, in order for genesis to function
properly. If you do not care what version is installed,
only that the executable exists, specify `null`, or `~` as
the version.

Example .genesis_deps file:

```
genesis: 1.5
spruce: 1.2.1
safe: ~
```