Env Hooks should be able to bail the creation step
If an environment hook (in .env_hooks) exits non-zero, Genesis now bails
and cleans up after the half-working environment that it tried to
create.  This allows our env hooks to be more assertive in their role as
configurators, and bail if they cannot continue (for example, if Vault
is not available).