Utilize the Genesis Index for Stemcell / Releases (#37)
* Utilize the Genesis Index for Stemcell / Releases

URLs, SHA1 checksums and versions (for "latest") will now be looked up
from the Genesis Index (per the GENESIS_INDEX environment variable).  By
default, this is set to https://genesis.starkandwayne.com/, but can be
overridden (for on-premise instances of the index) or disabled
completely by setting it to the value "no".

If a deployment supplies url file for a stemcell or a release, the trip
to the index will be skipped.  Any releases or stemcells that are
missing will consult the index.

This change also means that "latest" now no longer means "latest
available on your BOSH director", but instead means "latest known to the
index"

* Error out on missing SHA1 checksum file for Stemcells

* Access Genesis Index at time of `add release` / `use stemcell`

Instead of doing it at the metadata creation step (during a make
manifest), we access the Index when you add a new release to your
deployment, or use a stemcell.  This allows finer-grained control over
versioning, because you can just `add release x 1.2.3` when you need to.

* Refactor Genesis Index interaction

We now have a single call to the index, that occurs in a shared function
that handles both stemcell and release use cases.