Index-tracking 'track' Version Keyword
For stemcells and releases, the version keyword 'latest' is now back to
meaning "latest available on the director", and a new keyword, 'track'
can be used to track the latest available version in the Genesis Index.

This should make Genesis templates easier to write, since we can track
stemcells more easily, and people with good backwards-compatibility
discipline can even track releases.

bosh-init deployments are only allowed to specify explicit versions
(i.e. bosh 255.2) or the keyword "track", to track the Index.  The
"latest" keyword makes no sense in a bosh-init deployment, since we have
no BOSH director from which to get the latest release/stemcell.

Fixes #48