Fix an issue with `genesis embed` permissions (#53)
Previously, `genesis embed` would copy the source
file's permissions in place, if no genesis script
existed already in the deployment repo. When combined
with homebrew's genesis (permissions of 0555), this
meant subsequent `genesis embed`s would fail with
a `Permission Denied` error. Now, the embeded
copy is always 755 in the repo.