Add warning when using 'track' with CI pipelines.
Due to concerns with using 'track' in a pipelined Genesis
environment, or a production environment (#69), Genesis will
now issue a warning when trying to repipe a CI environment that
uses a release or stemcell with version 'track'.

If you wish to override this behavior, the GENESIS_CI_ALLOW_TRACK
environment variable must be set to 'true'.