Fix issue where stemcell/release version checking failed
The `--connect-timeout 1` flag on our `curl` calls to the BOSH director
to find existing release/stemcell information seems too agressive.
If there is a delay in the director responding to that request, it
caused `genesis` to try to upload the release/stemcell again, and fail.

Since prior to these `curls`, we have a failsafe switch checking to
ensure that the director is alive, the connect-timeout 1 doesn't buy us
anything but failures. If the director is down, the checks are
short-circuited entirely.