Bugfix: Better support in prompt_for_line for no prompt
By passing prompt of null to prompt_for_line, you can have the label as
the inline prompt, but this wouldn't work if there was also a default.
This fix puts the default at the end of the inline prompt instead of
awkwardly hovering above it with a single space indent.