chmod the params hook properly
The code for chmodding the params hook only did the chmod if the file
didn't exist (which is weird and wrong).  Fixing that led to the
discovery that we have a test that validates that the chmod was never
run (since the 0644 params hook _existed_).  This test has since been
removed.