Improve vault target handling
Test helper `vault_ok` now returns the target, which it passed to
t/bin/vault to create the target in ~/.saferc.  This allows the target
to be specified in one location should it ever need to change.

`teardown_vault` now unsets the `VAULT_PID` so that any call to vault_ok
won't erroneously state the vault was already started.