Top->download_kit no longer returns itself
This was done to improve output when `genesis download-kit` downloads
the latest version -- this now outputs the actual version semver of the
downloaded kit.  Further changes involved removing explains from Top and
putting them in the commands, as per the guiding priciples of keeping
the UI out of the OO components.

It did mean that the tests that chained the:
`my $top = Genesis::Top->create(...)->download_kit(...)`
had to be modified to make two calls.