Set BOSH director environment variables for hooks
When any hook (except prereqs and new) runs, it now gets the following
four environment variables, to make dealing with BOSH seamless:

  - BOSH_ENVIRONMENT   - Director URL
  - BOSH_CA_CERT       - Contents of the signing CA
  - BOSH_CLIENT        - Username logged into BOSH
  - BOSH_CLIENT_SECRET - Password for BOSH_CLIENT

This does rely on the user logging in with a client-analog account,
which is most likely going to be admin.  This is a stopgap / hack so
that we can write hooks that don't rely on `~/.bosh/config` and aliased
environments.  Eventually, Genesis will pull these four variables out of
Exodus data for the BOSH environment that deployed us.

Fixes #277