Added pre-deploy hook, pass data to post-deploy hook (#283)
The pre-deploy hook gets given the paths to an unredacted manifest and
a file that it can use for output.  This output filename is given to the
post-deploy hook if anything was written to the file.

The predeploy/postdeploy hooks handle the creation, reading and
deletion of the temporary data file themselves.  This limits exposure of
the file as it only exists for a short time.