Support Optional Parameters in kit.yml
The new `${maybe:param}` syntax allows Kit Authors to handle existence
and non-existence of parameters in the operator's environment file(s),
for reasons of certificate subject alternate name generation.

If the `maybe:`'d parameter doesn't exist in the environment file(s),
the empty string is returned (instead of signalling an error), and the
X.509 certificate translation to `safe` calls will omit that `--name`
flag.