cloud_config_needs static_ip ...
This adds the ability for kits to specify how many static_ips are needed, which is a better user experience than failing in spruce at manifest build time for each static IP that is missing.  Syntax is `cloud_config_needs static_ips <network_name> <quantity>`.  For example, the following directives from a mock cf-genesis-kit:

    cloud_config_needs static_ips $(lookup params.cf_internal_network cf-core) 15
    cloud_config_needs static_ips $(lookup params.cf_edge_network cf-edge) 25

might result in the following output:

    [my-env] running cloud-config and environmental parameter checks...
    Errors were encountered in your cloud-config:
     - network cf-core needs 15 static IP addresses, has 10
     - network cf-edge needs 25 static IP addresses, has 15
    ...

Like all cloud_config_needs call usage, this needs to be followed up with a block similar to the following to actually output the error messages and pass/fail:

    if check_cloud_config ; then
      describe "  cloud config [#G{OK}]"
    else
      describe "  cloud config [#R{FAILED}]"
      exit 1
    fi

Furthermore, the helper bash script that is used by the hooks is now checked with shellcheck, if present.  It has been added to the genesis Docker image in a previous direct-to-master commit.  It will fail on errors or warnings, but pass (and suppress) info and style level messages.  This is done to ensure any new potential errors are caught, but if it is deemed not an error, a directive can be added to the file to bypass that instance in the check (see https://github.com/starkandwayne/genesis/pull/297/files#diff-f57f19e6c402057bc4ea700b4515fdbaR417 for example)