Can ignore colour in templated csprintf output
It is often useful or needed to template the colour wrappers to use a
color that is specified in the arguments to csprintf, but sometimes that
color needs to be 'default' (i.e.: not specified).

By adding '-' to the valid colour codes to the recognized patern of
`#x{content}` where x is the colour code, the template can still be used
and not have to guess at what the default colour is.  When the colorizer
is called, it simply ignores it, but still removed the struture.

Also fixes a bug in the rainbow when $fg or $bg was not set (discovered
while testing this but not directly due to this)