Refactor Genesis::Vault for isolation
Genesis needs to be independent of the user's safe configurations for
multiple reasons, with the top two reasons being:
- Different users name their targets differently
- Prevent accidental changes to the wrong target

Furthermore, Genesis is inconsistant with its targeting of safe vaults.
Some commands, like new, explicitly request the user to target a vault,
changing the current target system-wide, and others just rely on the
desired vault already being the default target.

This refactor sets the groundwork for treating Vaults as first-class
citizens in the Genesis ecosystem, with each environment having a vault
that is specified by IP address and port.  Creation of new environments
prompts the user to specify the vault from a list of choices, and that
vault URL is added to the enviroment file under a new `genesis` top
level key, along with a verify TLS flag and secrets prefix.  Through
this, all further actions on that environment knows how to target the
desired vault.

Further commits will plumb this funtionality into the existing system,
including whatever tweaks are needed to support interface and file
changes.