Adds ability to use simple lable on selected choice
Genesis::UI::prompt_for_choice(s) routines allow alternate labels for
situations where the value is not human-readable or abstract.  This
change further allows differentiation between the label shown in the
choice list and the one printed on the 'select> ' line after a selection
has been made.  This is primarily used when the choice label is composed
of complex or tabular strings, and only a simple identifier is desired
on the select> line.