Allow empty string as a valid entry
Genesis::UI::prompt_for_line allows a blank line to be used to indicate
that the default value is desired.  However, this means that an empty
string cannot be entered.

This commit makes it so that if "" is specified as the default value, a
user can hit enter on blank entry to get an empty string as the value.
In this case the `(default: <value>)` label will not be shown.