Improve read_json_from
read_json_from now actually ignores the return code `rc` that is passed
in from the wrapped `run` call.  Now it will try to parse the output
regardless of the return code, then returns the json, rc and any
exception that may have been raised if the json could not be parsed.  If
the request is in scalar context, it only returns the json, and will die
if there was an error.