Don't hang while checking if host is reachable
In the `tcp_listening` function, it _should_ return immediately if the
port is not reachable, but under some circumstances, it just hangs.  To
work around this, we try to ping the host and time out after 10 seconds
if it doesn't respond.  If it does respond, it only adds 0.3seconds
to the round-trip ping time.