Finish renaming params.vault to secrets_path
Since we've added an object named Vault, the existing concept of the
vault path, known as vault_prefix, vault, and secrets_prefix in various
previous incarnations of this code, needs its own identity.

This updateis the instances to `secrets_path` in the code in all places
except where necessary for backwards compatibility of legacy
kits/environment files.