Improvements to debugging and help (#325)
* Improved dump_var output for run

The information that the run output comes from the run command is almost
never what is wanted, so dump_var now takes an optional prefix argument
of depth (0 is default) and run uses this to report the location of the
caller that resulted in the command output

* Add help line for secrets-provider

* Use better triangle for dev logs

The right triangle "" being used, while a better fit, is only supported
by specialized fonts (notably the powerline fonts).  Instead, "⮀" should
be used as it is supported by most standard modern fonts, although it is
slightly less tall.