Add `offer_environment_editor` helper with man support
Fixes #324

Creates a generic offer to edit generated environment file prior to
running the secrets generation.  On editors that support it, this will
also include the kits manual in a right-side split window.

Also adds documentation for the `offer_environment_editor` and the
`genesis_config_block` helper to the kit authoring new hook document.

Finally, the use of the deprecated GENESIS_VAULT_PREFIX is replaced with
GENESIS_SECRETS_PATH (GENESIS_VAULT_PREFIX will still get populated and
provided to hooks, but internally it now uses the newer
GENESIS_SECRETS_PATH)