Purge-secrets now manditory if secrets present
Bails on creation of new environment if there are secrets present and
the user elects not to purge them.

Not purging secrets and still continuing provided a false sense that
those secrets would still be used, but the new action overwrites any
existing secrets that it knows about, but leaves orphaned secrets from
previous versions of the kit, causing confusion when debugging.