Altered arguments due to `run` changes
`run` routine added a third value to its return for stderr.  This has
been extended to wrapper routines of `lines`, `read_json_from` and
`load_yaml_file` to support better handling of outputs.

The third argument of $err is passed down to the caller for
`read_json_from` and `load_yaml_file`, with some extra error handling
during the extra processing.

For `lines`, at this time, $err is dropped.  If extra processing is
required, we recommend separating the call to run and passing the
results to lines.