Fix test pollution for dev genesis versions
With the addition of checking kit prereqs, tests that run against
development versions (aka versions that don't have a release version)
caused extra noise in the tests complaining that the genesis could not
confirm that the version meets the needs of the test.  This has been
suppressed with the ability to be re-enabled with the environment
variable GENESIS_TESTING_DEV_VERSION_DETECTION when testing that
functionality explicitly.