Fix tar OS-interoperability issue on decompile
On MacOS, tar -t lists the files in the tarbal, which is used by Genesis
to determine that there was just a single directory.  On Linux, it also
lists the permission, owner and group (similar to ls -la).  This
different behaviour causes the compiled kit to be erroneously identified
as invalid.

By using awk to look at just the last column, we can harmonize the
output and resolve this inconsistancy.