Resolve UTF-8 by ignoring it
UTF-8 has been a plague on Genesis ever since trying to add it.  Too
many issues with different terminals misconfigured, and too much (or not
enough re-encoding and decoding).

To resolve this, we've gone to just using 'use utf8' in files that have
it, and suppressing any warnings'  If users still have a bad experience,
they can set $GENESIS_NO_UTF8 to 1 and it will degrade to ANSII
characters.

This also improves the Helper script `describe` by using the calling
Genesis' library instead of assuming its in $HOME/.geese/lib.

Furthermore, when dealing with invalid encoding, it now only prints out
one error instead of one for each argument.