Simplify glyphs, and allow bullet to decorate
The check, x and bullet ASCII representations were enveloped in
parentheses, which while providing a distinct look, resulted in an
undesired look when bullet routine had box on: `[(+)]` for check as an
example.

The outer parens were dropped fro consistent look and feel, and as they
are wrapped with `#@{?}`, there is no concern of a normal character
being accidentally converted to a UTF-8 glyph as was the case when this
was initially conceived.