Renamed _slice to get_opts, moved to lib/Genesis.pm
The poorly named _slice routine was used to slice out a subset of
options from the options hash, and optionally translate dashes to
underscores to make working with the keys easier.  This was a local
routine in Genesis::Env, but it is useful in general, especially in
bin/genesis, so has been extracted to the main utility library
lib/Genesis.