Added custom msg to die_unless_controlling_terminal
While its all good for developers of Genesis to know where in the code a
prompt was called in a non-controlling terminal that resulted in the run
being aborted, end users appreciate a more meaningful message.

Default is still where in the code the die_unless_controlling_terminal
was called.