Added warnings to secrets check
Some "failures" encountered are not errors, but warnings that things
aren't as exact as they could be.  Primarily is the usage checks in the
X509 secrets, especially the defaults for this.  In `safe v1.5.0`, the
default usage changed from nothing to `client_auth` and `server_auth`
for all certs, and additionally `crl_sign` and `key_cert_sign` for CA
certs.  Rather than have this a hard failure, it is now indicated as a
warning, and deployment can continue.

The --fail-on-warn option is provided to treat these warnings as errors
in the case that that behaviour is desired.

Addional to support of this feature, the working of failure checks when
validating was improved (found in check-secrets and done in
rotate/replace-secrets became valid), and a new glyph was added for
warnings.  `no-prompt` because `no_prompt` due to use of get_opts, and
boxes around bullets got the abiltiy to specify end glyphs as an array
(left-wall, right-wall).