Secrets prompt needs env before env exists
When in a `new` hook and prompting for secrets, the prompt routine needs
an Genesis::Env object to know where the secrets mounts are and if its
kit supports new vs old Genesis, and so on, but the env file hasn't been
created yet.

This solution creates a Genesis::Env object out of the environment
variables being passed back in from the hooks environment, which has all
the information needed to build a passable Env object for accessing kit
and vault activities.

It also fixes the call that was failing to create the needed
environment.

Fixes #389