Add scoping to the commands
Genesis commands can be scoped to their directory space.  There are
commands that don't need a specific scope, such as `help` and `version`,
and commands that need to be in a repo, or target a specific environment
file.  There are commands that CANNOT be in a repo too. There are even
commands that change scope depending on their options.

This change adds scoping to commands so that the scope can be checked
at start up, which removes the duplicate `check_prereqs` in every
command routine.  It also makes the global -C option work with
specifying the environment or not.

Specifically, this commit adds an optional `property` hash-ref after the
command name in the command arguments (defaults to an empty hash-ref),
and sets that as the $PROPS value for the command.  The first property
added is `scope`, which can be a single scalar, or an array ref.  If its
an array ref, each element can also be a scalar string or an array ref
of two array refs: the first being a list of option validations, and the
second being a list of valid scope strings.  If either of these are
single, it can be a single scalar value instead of an array ref.

The option validation can take the form of (!)<opt-name>(=<value>).  If
its just <opt-name>, it validates to the boolean value of the option.
If a leading `!` is present, it negates that validation.  If the
=<value> is present, it tests equality to the value instead of the
boolean value of the option.

When testing against option values, the first match defines the valid
selection of scopes.  For example, take this scope definition:

```
{scope => ['repo', ['dev' => 'repo'], 'kit']}
```

The `command_scope` command takes a scope label and options hash.  If
called, it would return true if the scope label given was 'repo' or
'kit' if there were not options, but if the option included a boolean
true for 'dev', 'kit' would no longer be valid because it comes after
the test for 'dev' specifies only 'repo' is valid.  However, repo is
valid whether 'dev' is true or not because its first in the list and
unqualified'.

Option validations in a single array-ref are and-ed together.  If the
scope was

```
[[['mode=new','!testing'] => ['prod','dev']],['mode=new' => 'qa'], 'sandbox']
```,

command_scope($x, mode => 'new', testing => 1) would be false prod, dev
and sandbox, becase it doesn't match the first criteria but does match
the second.