Better dev kit identity
While a dev kit's name and version are stuck being dev and latest for
legacy reasons, the dev kits id has been updated to reflect whatever
name and version are found in the dev kit's kit.yml, with a "(dev)" flag
tacked on the end.  If there is no name or version, the values of
"unknown" and "in-development" are used as filler respectively.  This
gives the user a better idea of what class of kit they are dealing with

In order to use this value in hooks, $GENESIS_KIT_ID was made available.