Add provided secrets handling
Users can now manage provided secrets via the *-secrets commands.
Provided secrets are those secrets that are not generated, but asked for
in the `genesis new` process.  A `genesis remove-secrets --all` will
remove them.  Copying an environment file to a new name and running
`add-secrets` wouldn't populate these, and `check-secrets` wouldn't
detect that they were missing.  That's no good!

This is now fixed.

* `check-secrets` now reports on their presence or absence
* `add-secrets` and `rotate-secrets` now propts for values to be stored.
* `remove-secrets` can now remove them

Support for this functionality must be provided per kit, which requires
a new top-level section in `kit.yml`: `provided`

The structure is thus:

```
provided:
  <"base"||feature-name>:
    <path>:
      [type: 'generic']
      keys:
        <key1>:
          [prompt: "informational discription of the source or purpose of the value"]
          [sensitive: true|false]
        [<key2>: ...]
    [<path2: ...]
  [<feature2>: ...]
```

`prompt` if not given defaults to "Value for <path> <key>" so its best to
provide one, but not fatal if you don't.

`sensitive` defaults to true - this means the input will be hidden and
confirmed with a second entry.  The user's input will be displayed if
false.

Also updated:

* Fixed filter bug when only paths were set.
* Removed stubbed "batch mode" for secrets processing, as it wouldn't
  work with 'provided' secrets.

* Version check for compile-kit to ensure this version of genesis is new
  enough to know how to compile a kit based on its `genesis_version_min`