Ensure genesis.env is present, warn on params.env
2.6.13 deprecated `params.env`, and for a brief time printed a warning
to that effect until it was determined that it was too noisy.

2.7.0-2.7.6 removed the usage of `params.env` and enforced the migration
of params.env to genesis.env, and kits declared with minimum version of
2.7.0 were expected to use `genesis.env` where they before used
`params.env`.

After pushback, we have decided to soffen the stance, and print warnings
when the environment uses a kit with genesis_version_min of 2.7.0 or
higher.  As these kits require genesis.env to be set, we do so as part
of the manifest merge.