Use internal connection timeout mechanism
Hopefully final salvo on the prevention of hanging connections to Vault
and BOSH.  The last version successfully used `timeout`, but this was
not as ubiquitous as we though, and most MacOS users were lacking this
utility.

We have now switched to using perl methods for timing out and checking
the availability of the target host:port instead of relying on external
utilities.

We have also clarified in our error messages if the connection attempt
timed out or simply failed.

Furthermore, we now check bosh connectivity in the helper scripts when
calling its `bosh` function, which prevents hook scripts from hanging
(in this case blueprint trying to find the bosh cpi).

Fixes #412