Vault now validates legacy connections
If using legacy mode, the vault specified would just be used without
validating that it was available.  This caused problems for obvious
reasons, but they would happen in the middle of operations and the
failure would be obscure.  Checking before hand makes this cleaner.

Also discovered a bug in the connect_and_validate routine that didn't
work if the secrets mount didn't have the handshake endpoint, so we
check both the specified secret mount as well as the default '/secret'
location for it.