Add test hook for check secrets validation bypass
By setting GENESIS_TESTING_CHECK_SECRETS_PRESENCE_ONLY to a truthy
value, `genesis check` will only check the presence and not the validity
of secrets.  While this is normally undesirable, it is necessary for
some testing, specifically the new testkit tool which standardizes the
secrets to non-random but invalid values.