Prevent bosh function calls
If your bosh command is named bosh, which most are these days, there was
an error that caused the bosh helper script to call back into Genesis to
check the bosh availability, which then caused Genesis to call bosh, but
the environment that loaded the helper script used the same name for
bosh, so it got called instead, which errored out because we clear the
bosh variables in the ping activity.

Now, when we initially determine the name of the bosh command, we do our
best to get its full path, or failing that, prefix it with a \ so that
it ignores functions and aliases.