Don't overwrite the current ENV for env vars
When Gensis goes to run hook (and now manifest creation), it sets up
environment variables.  Initially the hooks code created a local
variable for the duration of the hook run, but when the Genesis
environment's environment variables were extracted, it just wrote them
to the global scope, and when it was then used for manifest generation,
it ended up poluting the global scope.

It now just provides a map, which the calling function can inject into
its local copy of %ENV, or inject it straight into the envs section of
the run command.