Add features hook
While blueprint hook has the ability to make decisions on when a feature
is NOT present, or on specific combinations of features, that ability is
beyond other interactions.

We used to have a subkit hook which would allow you to create derived
features so that default features and not-features could show up as
explicit features, which allows things like secrets management to
determine dependencies for these.

This has been re-realized as a `features` hook, which given a list of
features in the `$GENESIS_REQUESTED_FEATURES` value, can provide a
derived list of features, which will be used by internal genesis for the
environment's features list, which in turn will be used to populate
`$GENESIS_REQUESTED_FEATURES` for other hooks.