Improvements on specifying configs.
The --cloud-config|--cc and --runtime-config|--rc have been streamlined
into a single --config|-c with backwards compatibility to the existing
-c (for cloud config).  You can now specify named configs as such:

```
-c [type[@name]=]/path/to/config.yml
```

If type is not given, it is assumed cloud, and likewise if name is not
given, it is assumed to be the unnamed `default` config for the given
type.

`-c` can be specified multiple times to specify multiple configs.  It
does not error check that you haven't specified the same type and name
multiple times, so that's on you to ensure you're not doing that.