Add better kit id to exodus data
As more things use the previously deployed kit to determine what needs
to be upgraded, it is important to have this information correctly
identified.  Prior to this change, dev kits reported the name as dev and
the version as latest (from the env yaml file)

This change uses the kit name and version located in the kit.yml file
directly, and also adds `kit_is_dev` to record that a dev kit is being
used.