More dynamic 'default' pipeline layout.
If ci.yml specifies a `default` layout, that layout will be used without
having to specify it, but it will expect the fly target to also be
`default` -- this is at odds with the concourse `login` addon which
names the fly target the same as the environment name.

To resolve this, if you only have a single layout in your `ci.yml` file,
it will be considered the default, so it can be named the same as your
fly target aka concourse environment.  If you have muptiple layouts, a
layout named `default`, if it exists, will be considered default to keep
existing behaviour.  Otherwise, if you have multiple named layouts and
you didn't specify one in the `repipe` command, it will present you with
a list to chose from.  The `-t|--target` command will still behave as
normal.