Enable Enterprise Vault support, use Safe to init
Instead of using the Vault Genesis kit, some clients need to integrate
with their companies Enterprise Vault.  This means supporting namespaces
and disabling strongbox, the process that Safe uses to treat multiple
Vault VMs as a single target when unsealing.

Since Safe nicely wraps up support for this, the pipelines have been
updated to use Safe to initialize connections to the Vault, making it
seamless regardless of it being Enterprise or not, v1 or v2 kv backend.

Changes:
  - In your `ci.yml`, under `pipeline.vault`, you can specify
    `namespace` as a string, and `no-strongbox` as a truthy value.