Add auto-update to pipeline
This creates an update-genesis group that contains a job that is
triggered when a new verison of the kit is released, and will update the
kit version in the deployment repo, and embed the latest version of
genesis if not at the latest version.

It will then commit these changes, which will trigger the primary
pipeline for the repo to progress through the various environments.

Supports separate kit and genesis access tokens (when kit provider is
not github.com)