Fix pipeline propagation of cached changes
Scenario this fixes:

Starting State
--------------

The following files exist:
- c.yml
- c-dev.yml
- c-qa.yml
- c-prod.yml

c-qa.yml was last updated, which triggered the changes input of the
c-qa deploy pipeline.

Issue
-----

- Update c.yml, and commit it
- c-dev deployment pipeline gets triggered - it has no previous, so it
  runs off its base environment files.
- c-dev generates cache by adding base c.yml and c-dev.yml to the
  `.genesis/cached/c-dev/` directory.
- c-qa pipeline is triggered by the 'c-qa-cached' input.
- c-qa correctly deployes based on c.yml from the cache directory and
  c-qa.yml from the base directory.

- ***ISSUE***: c-qa generates cache by adding c.yml and c-qa.yml to the
  `.genesis/cached/c-qa/` directory, NOT the files it deployed with.

- c-prod pipeline is triggered by the 'c-prod-cached' input.
- c-prod gets deployed with the old version of c.yml from the outdated
  'c-qa-changes' repo.

Resolution
----------

The generation of the cached files now takes the previous env and the
cache dir as input, and will propagate the version of the files used to
deploy to the next environment's upstream cache.