Hooks bail on missing SAFE_TARGET only if safe used
Some hooks don't need a safe target, or are run when the SAFE_TARGET
isn't available.  In these cases, we will override the safe call with a
function to bail if used.

This should be safe (no pun intended) for spruce because although spruce
doesn't use safe to interact with vault, it is also read-only, so risk
of damage is minimal.

It can still be bypassed if using `\safe` or `command safe`, so those
should be avoided in all hooks.