Remove kit hook, use cachable kit-overrides.yml file
An attempt was made to make secrets specified in the deployment repos to
support a 'generic' genesis kit.  In that case, the kit specified a
hooks/kit script that specified how to assemble the kit.yml file.

However, due to chicken/egg and cart/horse issues around compiling kits
and having a kit object for running the kit hook before loading the kit
definition, this approach failed. Furthermore, there are valid reasons
to override a kit configuration beyond the kits that would support it
with a hooks/kit.

Therefor the spike for the hooks/kit was reverted, and in its place is
support for the detection and merge of a ./kit-overrides.yml file.
Furthermore, if this file was cached from a previous environment (in the
case of pipelines), the cached value would be used.

It also allows for multi-doc support, and in theory, could be
dynamically created as part of the new, blueprint or addon scripts.