Ensure YAML partition is on a new line
YAML files are text files, and text files (in *nix) consist of a series
of lines, each terminated by a newline character, including the last
one.  So this is the behaviour Genesis expected when merging multiple
source YAML files into a multi-doc file.

HOWEVER, _certain_ editors don't naturally obey this spec, and when
Genesis encountered these files in the wild, the "---\n" separator was
appended to whatever the last value in the previous file was, which
didn't result in a proper multi-doc file and data was corrupted.