Fix unauthenticated BOSH report
The `bosh -e <env? env` command erroneously states you're not logged in
if your access token has expired but your refresh token is still valid.
The command also does not refresh your access token, but other
commands do, so we run a stemcell command to trigger the refresh and
then check env again to see if we are in fact logged in.