Reauth to vault right after deploy completes
While re-authenticating to vault before writing exodus data (added in
v2.7.22) solved some problems, there are a multitude of activities that
happen between the bosh deploy command completing and the writing of
exodus data that MAY try to interact with the Vault.  To resolve this,
re-authentication is performed immediately after the bosh deploy command
is complete.