Adds caching for ops/* and kit-overrrides.yml
Note:  This adds caching behaviour to the ops/* directory, ensuring that
the cached values are propagated from previous passed jobs in the
pipeline.  However, each kit will need to detect and make use of them in
its blueprint hook.

To update kits to do this, they will need to detect the presence of the
PREVIOUS_ENV environment variable, and then the presence of
.genesis/cached/$PREVIOUS_ENV/ops/<feature>.yml