Add support for multi-deployment repos in pipeline
In order to use pipelines, the .genesis directory and all the
environment files had to be at the base of your repo.  However, it is
common to have your different kit deployments in a single large repo.
This change makes it possible to support that configuration, with each
subdirectory having its own ci.yml that specifies the root subdirectory
for the pipeline.