Improve vault auto-authentication routines
Vault auth-authentication routines worked well, except for two issues:

1) If the vault being authenticated to was the vault being deployed, the
   post-deploy hooks would never run to unseal the vault because the
   vault couldn't be authenticated because it was sealed.

2) Checking the status of the vault automatically authenticated to the
   vault instead of telling the real status (internal) which caused
   failures.

Fix was to reorganize the process of post-deployment, to authenticate if
not sealed, with some fine tuning of preparing the cached manifest
before deploying when the vault is still available.

Also fixed was the Genesis::Vault#status command to not auto-
authenticate, while the external facing secrets-provider command will
still try to auto-authenticate for the user.